refactor(网络): network 路由下沉 service 层
- 新增 services/network_config.py：视图装配、保存编排（镜像 diff →
  重建媒体服务 → 闭合熔断）、连通性探测与结果分类全部下沉
- 请求/响应模型从路由文件迁至 schemas/network.py
- LLM 端点/密钥解析提为 llm_config.resolve_provider_endpoint 唯一判据，
  网络探测复用之，路由层不再接触 Repository 与密钥解密
- routes/network.py 由 320 行减至 59 行（纯参数进出）
